Decouple sandboxing utilities from application
Closes: https://github.com/metarhia/impress/issues/883
