+++ lib/application.js
@@ -67,7 +67,7 @@ const mixin = application => {
   }
 
   application.createApplication = callback => {
-    application.createSandbox(() => {
+    impress.createSandbox(application, () => {
       application.loadConfig(() => {
         application.loadDatabases(() => {
           application.loadPlaces(() => {
@@ -159,125 +159,6 @@ const mixin = application => {
     return USE_STRICT + code;
   };
 
-  // PrepareScript
-  //   fileName <string> file name (absolute path)
-  //   source <string> JavaScript code
-  // Returns: <Object> exported from script
-  application.prepareScript = (fileName, source) => {
-    try {
-      const key = application.relative(fileName);
-      const options = {
-        filename: fileName,
-        timeout: SCRIPT_PREPARE_TIMEOUT
-      };
-      const wrapper = key.startsWith('/api/') ? apiWrap : codeWrap;
-      const code = wrapper(source);
-      const script = new api.vm.Script(code, options);
-      const exports = script.runInContext(application.sandbox, options);
-      application.cache.scripts.set(key, exports);
-      return exports;
-    } catch (err) {
-      application.logException(err);
-      return null;
-    }
-  };
-
-  // Create script
-  //   fileName <string> absolute path
-  //   callback <Function>(err, exports)
-  application.createScript = (fileName, callback) => {
-    const key = application.relative(fileName);
-    let exports = application.cache.scripts.get(key);
-    if (exports) {
-      callback(null, exports);
-      return;
-    }
-    api.fs.readFile(fileName, (err, code) => {
-      if (err) {
-        impress.log.error(impress.CANT_READ_FILE + fileName);
-        callback(err);
-      } else {
-        exports = application.prepareScript(fileName, code);
-        callback(null, exports);
-      }
-    });
-  };
-
-  // Synchronous version of application.createScript
-  //   fileName <string> absolute path
-  // Returns: module exports
-  application.require = fileName => {
-    const key = application.relative(fileName);
-    let exports = application.cache.scripts.get(key);
-    if (!exports) {
-      try {
-        const code = api.fs.readFileSync(fileName);
-        if (!code) impress.log.error(impress.CANT_READ_FILE + fileName);
-        else exports = application.prepareScript(fileName, code);
-      } catch (err) {
-        application.logException(err);
-      }
-    }
-    return exports;
-  };
-
-  // Global require for sandbox
-  //   moduleName <string>
-  const globalRequire = moduleName => {
-    let exports;
-    if (!moduleName.includes('..')) {
-      const path = application.dir + '/node_modules/' + moduleName;
-      try {
-        exports = require(path);
-      } catch (err) {
-        application.logException(err);
-      }
-    } else {
-      application.logException(new Error(
-        `Access denied. Application can not require module: ${moduleName}`
-      ));
-    }
-    return exports;
-  };
-
-  application.createSandbox = callback => {
-    const sandbox = {
-      api: {},
-      console: api.con.createConsole(application.name)
-    };
-    sandbox.global = sandbox;
-    sandbox.application = application;
-    application.sandbox = api.vm.createContext(sandbox);
-    application.getSandboxConfig(() => {
-      const globals = application.config.sandbox.global || DEFAULT_SANDBOX;
-      for (let i = 0; i < globals.length; i++) {
-        const moduleName = globals[i];
-        let moduleLink;
-        if (moduleName === 'require') moduleLink = globalRequire;
-        else moduleLink = global[moduleName];
-        if (moduleLink) {
-          if (moduleName in api.registry.deprecated) {
-            const msg = api.registry.deprecated[moduleName];
-            moduleLink = api.registry.deprecate(moduleLink, msg);
-          }
-          application.sandbox[moduleName] = moduleLink;
-        }
-      }
-      const apis = application.config.sandbox.api || api.registry.defaultNames;
-      for (let j = 0; j < apis.length; j++) {
-        let moduleName = apis[j];
-        let moduleLink = api[moduleName];
-        if (!moduleLink) moduleLink = api.registry.require(moduleName);
-        if (moduleName === 'fs') {
-          moduleLink = api.sandboxedFs.bind(application.dir);
-        }
-        moduleName = api.common.spinalToCamel(moduleName);
-        if (moduleLink) application.sandbox.api[moduleName] = moduleLink;
-      }
-      callback();
-    });
-  };
-
   application.getSandboxConfig = callback => {
     if (application.isImpress) {
       application.config.sandbox = DEFAULT_API;
@@ -326,7 +207,7 @@ const mixin = application => {
               depth - 1, cb
             );
           } else if (fileExt === 'js') {
-            application.createScript(filePath, cb);
+            impress.createScript(application, filePath, cb);
           } else {
             cb();
           }
@@ -397,7 +278,7 @@ const mixin = application => {
             cb();
             return;
           }
-          application.createScript(filePath, (err, exports) => {
+          impress.createScript(application, filePath, (err, exports) => {
             if (err) {
               cb(err);
               return;
@@ -468,7 +349,7 @@ const mixin = application => {
       callback();
       return;
     }
-    application.createScript(configFile, (err, exports) => {
+    impress.createScript(application, configFile, (err, exports) => {
       if (err) {
         callback();
         return;
@@ -615,7 +496,7 @@ const mixin = application => {
   //   callback <Function> after file loaded
   application.loadPlaceFile = (placeName, path, file, callback) => {
     const sectionName = api.path.basename(file, '.js');
-    application.createScript(path + '/' + file, (err, exports) => {
+    impress.createScript(application, path + '/' + file, (err, exports) => {
       if (err) {
         callback();
         return;
