+++ lib/cache.js
@@ -51,7 +51,7 @@ const mixin = application => {
 
   extUpdate.js = (filePath, relPath) => {
     cache.scripts.del(relPath);
-    application.createScript(filePath, (err, exports) => {
+    impress.createScript(application, filePath, (err, exports) => {
       if (err) return;
       cache.scripts.add(relPath, exports);
       const sectionName = api.path.basename(filePath, '.js');
