+++ lib/client.js
@@ -764,7 +764,7 @@ class Client {
   runScript(handler, fileName, callback) {
     const application = this.application;
 
-    application.createScript(fileName, (err, fn) => {
+    impress.createScript(application, fileName, (err, fn) => {
       if (err) {
         this.err = err;
         callback();
