+++ lib/core.js
@@ -23,12 +23,8 @@ impress.nextWorkerId = 1;
 impress.applications = {};
 impress.stat = { fork: 0, event: 0, req: 0, res: 0 };
 
-require('./constants');
-require('./stack');
-require('./plugins');
-require('./workers');
-require('./files');
-require('./templates');
+['constants', 'stack', 'plugins', 'scripts', 'workers', 'files', 'templates']
+  .forEach(name => require('./' + name));
 
 impress.serverConfigDefinition = api.definition.require(
   'config.impress.definition'
@@ -57,7 +53,7 @@ impress.fatalError = err => {
   }
   impress.log.error(msg);
 
-  if (msg.includes('application.createScript')) {
+  if (msg.includes('impress.createScript')) {
     impress.log.system(
       'Recover worker after throw Error in application: ' + appName
     );
@@ -210,7 +206,7 @@ const loadApplications = callback => {
 
 impress.start = () => {
   impress.loadPlugins();
-  impress.createSandbox(() => {
+  impress.createSandbox(impress, () => {
     impress.loadConfig(() => {
       if (impress.workerType === 'long') {
         process.title = 'impress ' + impress.nodeId;
